+++ main.py
@@ -1,6 +1,7 @@
 from flask import Flask, request, redirect, render_template
 from flask_sqlalchemy import SQLAlchemy
 from datetime import datetime
+import random
 
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///test.db'
@@ -8,23 +9,21 @@ db = SQLAlchemy(app)
 
 class Thought(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    username = db.Column(db.String(20))
     content = db.Column(db.String(140), nullable=False)
     date_created = db.Column(db.DateTime, default=datetime.utcnow)
 
     def __repr__(self):
         return f"<Thought {self.id}>"
 
-db.create_all()
-
 @app.route('/')
 def index():
-    return render_template('index.html')
+    thoughts = Thought.query.order_by(Thought.date_created).all()
+    return render_template('index.html', thoughts=thoughts)
 
 @app.route('/new', methods=['POST'])
 def new():
     if request.method == "POST":
-        thought = request.form['thought']
+        thought = request.form['name']
         new_thought = Thought(content=thought)
 
         try:
@@ -34,11 +33,16 @@ def new():
         except:
             return 'Error adding thought'
 
-        
-@app.route('/show')
+@app.route('/viewall')
 def view_all():
     thoughts = Thought.query.order_by(Thought.date_created).all()
-    return render_template('show.html', thoughts=thoughts)
+    return render_template('all_thoughts.html', thoughts=thoughts)
+
+@app.route('/getrandom')
+def get_random():
+    rand = random.randrange(0, Thought.query(Thought).count()) 
+    thought = Thought.query(Thought)[rand].content
+    return render_template('random_thought.html', content = thought)
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    app.run()
